backend/requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.34,<1
python-multipart>=0.0.20,<1
sqlalchemy[asyncio]>=2.0,<3
asyncpg>=0.30,<1
alembic>=1.14,<2
pydantic-settings>=2.7,<3
bcrypt>=4.2,<6
python-jose[cryptography]>=3.3,<4
aioboto3>=13.0,<14
